Fix Index typehints and move class (#1838)
