+++ src/snowflake/snowpark/modin/plugin/extensions/index.py
@@ -27,11 +27,15 @@ from typing import Any, Callable, Hashable, Iterator, Literal
 
 import numpy as np
 import pandas as native_pd
-from pandas._typing import ArrayLike, DtypeObj, NaPosition, Self
+from pandas._typing import ArrayLike, DtypeObj, NaPosition
 from pandas.core.arrays import ExtensionArray
 from pandas.core.dtypes.base import ExtensionDtype
 
+from snowflake.snowpark.modin.pandas import DataFrame, Series
 from snowflake.snowpark.modin.pandas.utils import try_convert_index_to_native
+from snowflake.snowpark.modin.plugin.compiler.snowflake_query_compiler import (
+    SnowflakeQueryCompiler,
+)
 from snowflake.snowpark.modin.plugin.utils.error_message import (
     ErrorMessage,
     index_not_implemented,
@@ -39,11 +43,66 @@ from snowflake.snowpark.modin.plugin.utils.error_message import (
 from snowflake.snowpark.modin.plugin.utils.warning_message import WarningMessage
 
 
+def is_lazy_check(func: Callable) -> Callable:
+    """
+    Decorator method for separating function calls for lazy indexes and non-lazy (column) indexes
+    """
+
+    def check_lazy(*args: Any, **kwargs: Any) -> Any:
+        func_name = func.__name__
+
+        # If the index is lazy, call the method and return
+        if args[0].is_lazy:
+            returned_value = func(*args, **kwargs)
+            return returned_value
+        else:
+            # If the index is not lazy, get the cached native index and call the function
+            native_index = args[0]._index
+            native_func = getattr(native_index, func_name)
+
+            # If the function is a property, we will get a non-callable, so we just return it
+            # Examples of this are values or dtype
+            if not callable(native_func):
+                return native_func
+
+            # Remove the first argument in args, because it is `self` and we don't need it
+            args = args[1:]
+            args = tuple(try_convert_index_to_native(a) for a in args)
+            for k, v in kwargs.items():
+                kwargs[k] = try_convert_index_to_native(v)
+            returned_value = native_func(*args, **kwargs)
+
+            # If we return a native Index, we need to convert this to a modin index but keep it locally.
+            # Examples of this are `astype` and `copy`
+            if isinstance(returned_value, native_pd.Index):
+                returned_value = Index(returned_value, convert_to_lazy=False)
+            # Some methods also return a tuple with a pandas Index, so convert the tuple's first item to a modin Index
+            # Examples of this are `_get_indexer_strict` and `sort_values`
+            elif isinstance(returned_value, tuple) and isinstance(
+                returned_value[0], native_pd.Index
+            ):
+                returned_value = (
+                    Index(returned_value[0], convert_to_lazy=False),
+                    returned_value[1],
+                )
+            # For methods that return a series, convert this series to snowpark pandas
+            # an example is to_series
+            elif isinstance(returned_value, native_pd.Series):
+                returned_value = Series(returned_value)
+
+            # for methods that return a dataframe, convert this dataframe to snowpark pandas
+            elif isinstance(returned_value, native_pd.DataFrame):
+                returned_value = DataFrame(returned_value)
+
+            return returned_value
+
+    return check_lazy
+
+
 class Index:
     def __init__(
         self,
-        # TODO: SNOW-1481037 : Fix typehints for index constructor, set_query_compiler and set_local_index
-        data: ArrayLike | Any = None,
+        data: ArrayLike | SnowflakeQueryCompiler | None = None,
         dtype: str | np.dtype | ExtensionDtype | None = None,
         copy: bool = False,
         name: object = None,
@@ -109,8 +168,7 @@ class Index:
 
     def set_query_compiler(
         self,
-        # TODO: SNOW-1481037 : Fix typehints for index constructor, set_query_compiler and set_local_index
-        data: ArrayLike | Any = None,
+        data: ArrayLike | SnowflakeQueryCompiler | None = None,
         dtype: str | np.dtype | ExtensionDtype | None = None,
         copy: bool = False,
         name: object = None,
@@ -119,11 +177,6 @@ class Index:
         """
         Helper method to find and save query compiler when index should be lazy
         """
-        from snowflake.snowpark.modin.pandas.dataframe import DataFrame
-        from snowflake.snowpark.modin.plugin.compiler.snowflake_query_compiler import (
-            SnowflakeQueryCompiler,
-        )
-
         if isinstance(data, SnowflakeQueryCompiler):
             qc = data
         else:
@@ -140,8 +193,7 @@ class Index:
 
     def set_local_index(
         self,
-        # TODO: SNOW-1481037 : Fix typehints for index constructor, set_query_compiler and set_local_index
-        data: ArrayLike | Any = None,
+        data: ArrayLike | SnowflakeQueryCompiler | None = None,
         dtype: str | np.dtype | ExtensionDtype | None = None,
         copy: bool = False,
         name: object = None,
@@ -150,10 +202,6 @@ class Index:
         """
         Helper method to create and save local index when index should not be lazy
         """
-        from snowflake.snowpark.modin.plugin.compiler.snowflake_query_compiler import (
-            SnowflakeQueryCompiler,
-        )
-
         if isinstance(data, SnowflakeQueryCompiler):
             index = data._modin_frame.index_columns_pandas_index
         else:
@@ -165,65 +213,6 @@ class Index:
                 tupleize_cols=tupleize_cols,
             )
         self._index = index
-
-    def is_lazy_check(func: Any) -> Any:
-        """
-        Decorator method for separating function calls for lazy indexes and non-lazy (column) indexes
-        """
-
-        def check_lazy(*args: Any, **kwargs: Any) -> Any:
-            func_name = func.__name__
-
-            # If the index is lazy, call the method and return
-            if args[0].is_lazy:
-                returned_value = func(*args, **kwargs)
-                return returned_value
-            else:
-                # If the index is not lazy, get the cached native index and call the function
-                native_index = args[0]._index
-                native_func = getattr(native_index, func_name)
-
-                # If the function is a property, we will get a non-callable, so we just return it
-                # Examples of this are values or dtype
-                if not callable(native_func):
-                    return native_func
-
-                # Remove the first argument in args, because it is `self` and we don't need it
-                args = args[1:]
-                args = tuple(try_convert_index_to_native(a) for a in args)
-                for k, v in kwargs.items():
-                    kwargs[k] = try_convert_index_to_native(v)
-                returned_value = native_func(*args, **kwargs)
-
-                # If we return a native Index, we need to convert this to a modin index but keep it locally.
-                # Examples of this are `astype` and `copy`
-                if isinstance(returned_value, native_pd.Index):
-                    returned_value = Index(returned_value, convert_to_lazy=False)
-                # Some methods also return a tuple with a pandas Index, so convert the tuple's first item to a modin Index
-                # Examples of this are `_get_indexer_strict` and `sort_values`
-                elif isinstance(returned_value, tuple) and isinstance(
-                    returned_value[0], native_pd.Index
-                ):
-                    returned_value = (
-                        Index(returned_value[0], convert_to_lazy=False),
-                        returned_value[1],
-                    )
-                # For methods that return a series, convert this series to snowpark pandas
-                # an example is to_series
-                elif isinstance(returned_value, native_pd.Series):
-                    from snowflake.snowpark.modin.pandas import Series
-
-                    returned_value = Series(returned_value)
-
-                # for methods that return a dataframe, convert this dataframe to snowpark pandas
-                elif isinstance(returned_value, native_pd.DataFrame):
-                    from snowflake.snowpark.modin.pandas import DataFrame
-
-                    returned_value = DataFrame(returned_value)
-
-                return returned_value
-
-        return check_lazy
 
     def __getattr__(self, key: str) -> Any:
         """
@@ -580,7 +569,7 @@ class Index:
 
     def set_names(
         self, names: Any, level: Any = None, inplace: bool = False
-    ) -> Self | None:
+    ) -> Index | None:
         """
         Set Index name.
 
@@ -945,7 +934,7 @@ class Index:
         # TODO: SNOW-1458147 implement drop_duplicates
 
     @is_lazy_check
-    def duplicated(self, keep: Literal["first", "last", False] = "first") -> Any:
+    def duplicated(self, keep: Literal["first", "last", False] = "first") -> np.ndarray:
         """
         Indicate duplicate index values.
 
@@ -1516,11 +1505,8 @@ class Index:
 
     @is_lazy_check
     def to_series(
-        self,
-        index: Index | None = None,
-        name: Hashable | None = None
-        # TODO: SNOW-1481037 : Fix typehints
-    ) -> Any:
+        self, index: Index | None = None, name: Hashable | None = None
+    ) -> Series:
         """
         Create a Series with both index and values equal to the index keys.
 
@@ -1544,8 +1530,6 @@ class Index:
         Index.to_frame : Convert an Index to a DataFrame.
         Series.to_frame : Convert Series to DataFrame.
         """
-        from snowflake.snowpark.modin.pandas import Series
-
         # get the index name if the name is not given
         if name is None:
             name = self.name
@@ -1564,8 +1548,7 @@ class Index:
         return ser
 
     @is_lazy_check
-    # TODO: SNOW-1481037 : Fix typehints
-    def to_frame(self, index: bool = True, name: Hashable | None = None) -> Any:
+    def to_frame(self, index: bool = True, name: Hashable | None = None) -> DataFrame:
         """
         Create a DataFrame with a column containing the Index.
 
@@ -1588,8 +1571,6 @@ class Index:
         Index.to_series : Convert an Index to a Series.
         Series.to_frame : Convert Series to DataFrame.
         """
-        from snowflake.snowpark.modin.pandas import DataFrame
-
         # Do a reset index to convert the index column to a data column,
         # the index column becomes the pandas default index of row position
         # Example:
@@ -2291,7 +2272,7 @@ class Index:
         raise TypeError("Index does not support mutable operations")
 
     @property
-    def str(self) -> str:
+    def str(self) -> native_pd.core.strings.accessor.StringMethods:
         """
         Vectorized string functions for Series and Index.
